requirements.txt
Django>=4.0.5
djangorestframework>=3.13.1
djangorestframework-simplejwt>=5.2.0
django-use-email-as-username>=1.2.0
django-cors-headers>=3.13.0
psycopg2-binary>=2.9.3
gunicorn>=20.1.0
requests>=2.25.1
Pillow>=9.0.1
drf-yasg>=1.21.3
redis>=4.3.4
celery>=5.2.7
tifffile>=2022.8.12
django-debug-toolbar
django-filter
django-prometheus
pynvml
psutil
Pydicom
markupsafe==2.0.1
